Attempt to make reopen-log work the Unix way
reopen-log must just re-open the log file(s)
and not attempt to do anything beyond this with them:
neither to rename them nor to rewrite.

This commit fixes the current ejabberd behavior, but
the old SASL log is still open after a call to
`ejabberdctl reopen-log`.
What log file is used for logging in this case is still unclean.