Explicitly set umask to 027 before starting erl
The previous value of 002 did not buy much as this is possibly
what erl have seen at startup anyway.
umask 027 revokes g=w,o=rwx from created files and dirs,
which means ejabberd creates files with rw-r---- and
dir with rwxr-x---. This seems to be OK for Mnesia files
and logs, but this does slightly change the behaviour of the
`ejabberdctl export2odbc` and similar commands which write
some data out, as this data is no more readable and writable
by the users other than ejabberd and root and those which are
members of the group ejabberd.
On the other hand, ejabberdctl commands which export DB data
are more secure as no regular user can now read this data,
and since they're supposed to be run by the root anyway,
this appears to be not bad.