Make backslash handling resemble sh_to_awk()
xmerl_regex:sh_to_awk/1 doubles *any* backslashes
it finds outside definitions of character classes ([...] groups)
and do not touch backslashes inside them.
We now implement the behaviour close to sh_to_awk;
the difference is that sh_to_awk uses rather simple-minded
approach to handling character classes: for instance, unclosed
char. classes are not detected and the case of a backslash
just before ']' is not detected as well.