Add simple test suite
Exports test:test/0 which just runs
xmerl_regexp:sh_to_awk/1 and ejabberd_regexp:sh_to_awk/1
on the same pattern and then uses the "result specification"
for that pattern to decide whether the test passed or not.

There are semantic differences between our and the reference
implementations though; test suite highlights them but it's
not clear how to handle them in a sensible manner.